Enable display of a collection specification on the details page (#1067)
* Enable display of collection spec on details page

* Add outlined toggle button group styling to theme

* Display loading state for collection spec views

* Convert icon outlined toggle button group to a text variant

* Move KeyAutoComplete header into independent file

* Correct typo in bindings editor store state action
