+++ src/components/tables/cells/fieldSelection/FieldActions.tsx
@@ -5,6 +5,7 @@ import {
     TranslatedConstraint,
 } from 'components/editor/Bindings/FieldSelection/types';
 import FieldActionButton from 'components/tables/cells/fieldSelection/FieldActionButton';
+import { outlinedToggleButtonGroupStyling } from 'context/Theme';
 import { useMemo } from 'react';
 import {
     useBinding_recommendFields,
@@ -69,25 +70,7 @@ function FieldActions({ bindingUUID, field, constraint }: Props) {
             <ToggleButtonGroup
                 size="small"
                 exclusive
-                sx={{
-                    '& .MuiToggleButton-root': {
-                        '&:not(:first-of-type), &:not(:last-of-type)': {
-                            borderRadius: 0,
-                        },
-                        '&:first-of-type': {
-                            borderTopLeftRadius: 4,
-                            borderBottomLeftRadius: 4,
-                            borderTopRightRadius: 0,
-                            borderBottomRightRadius: 0,
-                        },
-                        '&:last-of-type': {
-                            borderTopLeftRadius: 0,
-                            borderBottomLeftRadius: 0,
-                            borderTopRightRadius: 4,
-                            borderBottomRightRadius: 4,
-                        },
-                    },
-                }}
+                sx={outlinedToggleButtonGroupStyling}
             >
                 <FieldActionButton
                     messageId="fieldSelection.table.cta.includeField"
